FEAT: Call controller methods through all layers.
The main layers of any feature are application, domain and infrastructure, each one having a defined responsibility and approach for processing data.

To keep things simple and because this project is not big, mappers won't be used.

The focus will be on Hexagonal Architecture with vertical slicing and unit testing.