FIX: Change output dir.
It seems that this may be the problem with the 404 error on Netlify and Vercel.

Until now I managed to deploy this same project in another repo but having issues with this repo so I just used the same nuxt.config.js as it is the only thing I changed there.