FEAT: Complete basic cookie auth flow. Config notifications. Modify signin.
The auth flow was done this way to not require an extra Redis server, HTTPOnly cookies are more secure but they gave trouble without an external system due to the client not being able to get their value.

I didn't used localStorage because I was practicing auth with cookies and without abstraction layers to learn them better.

Finally the idea to save cookies into the Pinia store and read them from the server was promising but didn't work well, may be possible but not trivial to use Pinia+Nuxt 3+SSR